Move Bounds into its own source file, refactor soon.
