+++ bounds.go
@@ -1,0 +1,193 @@
+package btrie
+
+import (
+	"bytes"
+	"fmt"
+	"math"
+)
+
+// Bounds is the argument type for [OrderedBytesMap.Range].
+// A nil value for [Bounds.Begin] or [Bounds.End] represents +/-Inf;
+// which one depends on the value of [Bounds.Reverse].
+// Note that an empty bounds value is not nil; -Inf < []byte{} < []byte{0}.
+// For non-nil values, Begin is inclusive and End is exclusive regardless of the direction.
+// While a Bounds instance can be constructed directly, using [From] with [Bounds.To] or [Bounds.DownTo] is clearer.
+type Bounds struct {
+	Begin, End []byte
+	Reverse    bool
+}
+
+// From creates a new Bounds with bounds.Begin=key.
+func From(key []byte) *Bounds {
+	return &Bounds{key, nil, false}
+}
+
+// Clone returns a deep copy of b.
+func (b *Bounds) Clone() *Bounds {
+	return &Bounds{
+		bytes.Clone(b.Begin),
+		bytes.Clone(b.End),
+		b.Reverse,
+	}
+}
+
+func (b *Bounds) String() string {
+	if b.Reverse {
+		return fmt.Sprintf("[%X down to %X]", b.Begin, b.End)
+	}
+	return fmt.Sprintf("[%X to %X]", b.Begin, b.End)
+}
+
+// To sets b.End to key and b.Reverse to false, returning b.
+// To will panic if b.Begin >= b.End.
+func (b *Bounds) To(key []byte) *Bounds {
+	if b.Begin != nil && key != nil && bytes.Compare(b.Begin, key) >= 0 {
+		panic("bounds From >= To")
+	}
+	b.End = key
+	b.Reverse = false
+	return b
+}
+
+// DownTo sets b.End to key and b.Reverse to true, returning b.
+// DownTo will panic if b.Begin <= b.End.
+func (b *Bounds) DownTo(key []byte) *Bounds {
+	if b.Begin != nil && key != nil && bytes.Compare(b.Begin, key) <= 0 {
+		panic("bounds From <= DownTo")
+	}
+	b.End = key
+	b.Reverse = true
+	return b
+}
+
+// Compare returns where key is in relation to b, taking b.Reverse into account.
+// Compare returns 0 if within the bounds, -1 if beyond b.Begin, and +1 if beyond b.End.
+// Compare will panic if key is nil.
+// -Inf < {} < {0}.
+func (b *Bounds) Compare(key []byte) int {
+	if key == nil {
+		panic("key cannot be nil")
+	}
+	if b.Reverse {
+		if b.Begin != nil && bytes.Compare(key, b.Begin) > 0 {
+			return -1
+		}
+		if b.End != nil && bytes.Compare(b.End, key) >= 0 {
+			return +1
+		}
+		// end < key <= begin
+		return 0
+	}
+	if b.Begin != nil && bytes.Compare(key, b.Begin) < 0 {
+		return -1
+	}
+	if b.End != nil && bytes.Compare(b.End, key) <= 0 {
+		return +1
+	}
+	// begin <= key < end
+	return 0
+}
+
+// childBounds returns the start and stop key bytes, inclusive,
+// for the children of partialKey that a traversal should recurse into.
+// If b.Reverse is false or true, returns start <= stop or start >= stop respectively.
+// If ok is false, no children should be recursed into.
+//
+// For example, with partialKey {5, 8} and bounds [{5, 8, 4, 255} to {5, 8, 7}], return (4, 6, true).
+//
+//nolint:nonamedreturns
+func (b *Bounds) childBounds(partialKey []byte) (start, stop byte, ok bool) {
+	if b.Reverse {
+		return b.reverseChildBounds(partialKey)
+	}
+	return b.forwardChildBounds(partialKey)
+}
+
+//nolint:nonamedreturns
+func (b *Bounds) forwardChildBounds(partialKey []byte) (start, stop byte, ok bool) {
+	start, stop = 0, math.MaxUint8
+	keySize := len(partialKey)
+	if b.Begin != nil {
+		beginPrefix := b.Begin
+		if len(beginPrefix) > keySize {
+			beginPrefix = beginPrefix[:keySize]
+		}
+		switch bytes.Compare(partialKey, beginPrefix) {
+		case -1:
+			return 0, 0, false
+		case +1:
+			// start = 0
+		default:
+			if len(b.Begin) > keySize {
+				start = b.Begin[keySize]
+			}
+			// else start = 0
+		}
+	}
+	if b.End != nil {
+		endPrefix := b.End
+		if len(endPrefix) > keySize {
+			endPrefix = endPrefix[:keySize]
+		}
+		switch bytes.Compare(partialKey, endPrefix) {
+		case -1:
+			// stop = 0xFF
+		case +1:
+			return 0, 0, false
+		default:
+			if len(b.End) == keySize {
+				return 0, 0, false
+			}
+			stop = b.End[keySize]
+			// stop is inclusive, b.End is not
+			if len(b.End) == keySize+1 {
+				if stop == 0 {
+					return 0, 0, false
+				}
+				stop--
+			}
+		}
+	}
+	return start, stop, true
+}
+
+//nolint:nonamedreturns
+func (b *Bounds) reverseChildBounds(partialKey []byte) (start, stop byte, ok bool) {
+	start, stop = math.MaxUint8, 0
+	keySize := len(partialKey)
+	if b.Begin != nil {
+		beginPrefix := b.Begin
+		if len(beginPrefix) > keySize {
+			beginPrefix = beginPrefix[:keySize]
+		}
+		switch bytes.Compare(partialKey, beginPrefix) {
+		case -1:
+			// start = 0xFF
+		case +1:
+			return 0, 0, false
+		default:
+			if len(b.Begin) == keySize {
+				return 0, 0, false
+			}
+			start = b.Begin[keySize]
+		}
+	}
+	if b.End != nil {
+		endPrefix := b.End
+		if len(endPrefix) > keySize {
+			endPrefix = endPrefix[:keySize]
+		}
+		switch bytes.Compare(partialKey, endPrefix) {
+		case -1:
+			return 0, 0, false
+		case +1:
+			// stop = 0
+		default:
+			if len(b.End) > keySize {
+				stop = b.End[keySize]
+			}
+			// else stop = 0
+		}
+	}
+	return start, stop, true
+}
